requirement.txt
rich
websocket-client